+++ crates/types/src/structs.rs
@@ -2,6 +2,7 @@ use crate::SchemaField;
 
 #[derive(Clone, Debug, Default)]
 pub struct StructType {
+    pub description: Option<String>,
     pub fields: Vec<SchemaField>,
     pub name: Option<String>,
     pub partial: bool,